Split fanout tests off from general test_sequencer tests
